snapshot of Ant3.1b, more or less
some changes to make work on the current version of MacOS.

ports to Windows and older systems (BSDI, Solaris, etc) are
likely to be completely broken.  Some features of the GUI
do not work correctly, but the commandline tools seem to
work, at least for basic things.  The documentation builds
with current texlive releases.  My mac is ancient, and running
an OS several releases behind current, so it might not be
portable to any up-to-date OSs, except possibly Linux.